Login and SignUp moved to the correct folders
