--- client/src/pages/__pages.scss
@@ -1,1 +1,0 @@
-@import './_login.scss'
