--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "axios": "^1.1.3"
-  }
-}
